requirements.txt
appdirs==1.4.0
cffi==1.9.1
click==6.7
cryptography==1.7.2
Flask==0.12
idna==2.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyasn1==0.2.2
pycparser==2.17
PyJWT==1.4.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
Werkzeug==0.11.15
